M2 point 0 Group A: widen 44 int*/long* out-param locals to match callee's typed long* signature (13 root shapes, 18 files); fix 4 pre-existing %d/long format-string mismatches found along the way
FossilOrigin-Name: d8af32c158ce0ed9e8ed2db59ba0bfba30136ce1f6672f109ae885db99255f27

+++ src/atk/adew/cel.c
@@ -392,7 +392,7 @@ long cel__ReadFile(self,thisFile)
 struct cel *self;
 FILE *thisFile;
 {  
-    int objectID;
+    long objectID;
     long result;
     char *objectName;
     objectName = filetype_Lookup(thisFile, NULL, &objectID, NULL); /* For now, ignore attributes. */
